Terminado agregado y detalle de documentos

+++ src/main/java/me/hsanchez/digital_library/handlers/SearchResultHandler.java
@@ -1,5 +1,6 @@
 package me.hsanchez.digital_library.handlers;
 
+import java.math.BigInteger;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
@@ -22,7 +23,7 @@ public class SearchResultHandler extends BeanListHandler<DocumentDTO> {
 		
 		while(rs.next()) {
 			DocumentDTO document = new DocumentDTO();
-			document.setId(rs.getInt("id"));
+			document.setId(BigInteger.valueOf(rs.getLong("id")));
 			document.setTitle(rs.getString("title"));
 			document.setPrice(rs.getDouble("price"));
